fixed the url for the GET request from the game list and fixed the url implemented into the game list items so that the navigate to game detail screen. FINISHED FIRST SPRINT

+++ app/src/main/java/com/naughtsmt/lintu/data/data_source/dto/Msrp.kt
@@ -2,5 +2,5 @@ package com.naughtsmt.lintu.data.data_source.dto
 
 data class Msrp(
     val country: String,
-    val price: Int
+    val price: Double
 )
